requirements.txt

fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
pypdf
pillow
pytesseract
gradio
requests
